Move plot_api
As plot_api is used by both ert.gui.plottery and ert.gui.tools.plot,
its current location causes circular dependency problems. Moving
it to ert.gui resolves that problem.
