+++ src/ert/gui/plot_api.py
