+++ src/ert/gui/plottery/plots/ensemble.py
@@ -16,8 +16,8 @@ if TYPE_CHECKING:
     from matplotlib.axes import Axes
     from matplotlib.figure import Figure
 
+    from ert.gui.plot_api import EnsembleObject
     from ert.gui.plottery import PlotConfig, PlotContext
-    from ert.gui.tools.plot.plot_api import EnsembleObject
 
 
 class EnsemblePlot:
