+++ src/ert/gui/plottery/plots/gaussian_kde.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from scipy.stats import gaussian_kde
 
-from ert.gui.tools.plot.plot_api import EnsembleObject
+from ert.gui.plot_api import EnsembleObject
 
 from .plot_tools import PlotTools
