+++ src/ert/gui/tools/plot/filterable_kw_list_model.py
@@ -1,7 +1,7 @@
 from typing import Dict, List
 
 from ert.gui.ertwidgets.models.selectable_list_model import SelectableListModel
-from ert.gui.tools.plot.plot_api import PlotApiKeyDefinition
+from ert.gui.plot_api import PlotApiKeyDefinition
 
 
 class FilterableKwListModel(SelectableListModel):
