+++ src/ert/gui/tools/plot/plot_ensemble_selection_widget.py
@@ -21,7 +21,7 @@ from qtpy.QtWidgets import (
     QWidget,
 )
 
-from ert.gui.tools.plot.plot_api import EnsembleObject
+from ert.gui.plot_api import EnsembleObject
 
 
 class EnsembleSelectionWidget(QWidget):
