+++ tests/ert/unit_tests/gui/plottery/test_stddev_plot.py
@@ -5,9 +5,9 @@ import numpy as np
 import pytest
 from matplotlib.figure import Figure
 
+from ert.gui.plot_api import EnsembleObject
 from ert.gui.plottery import PlotConfig, PlotContext
 from ert.gui.plottery.plots.std_dev import StdDevPlot
-from ert.gui.tools.plot.plot_api import EnsembleObject
 
 
 @pytest.fixture()
